Add env var to allow keeping the existent navigation on a fragment load
Fragment navigation should be cancelled if there is ongoing navigation

Fragment navigation is synchronous, while navigation is asynchronous.
It can happen that fragment navigation starts when navigation is in
the policy or provisional step of loading. In that case, we should
cancel the requested fragment navigation and make the navigation finish.