Do not try to swap navigation as we're not using PSON
According to the specs, in order to enable SharedArrayBuffer we need to
set the appropriate network headers. When those headers are set, a new
web process in isolated mode is created, which is the one that will run
with SharedArrayBuffer enabled to avoid security problems.

But we're not using PSON here, so we don't ever want a new web process
to be created. To avoid this, we just force that any possible decission
that may create a new web process to return the value of not creating it.