Build a bundled copy of muprof https://bugs.webkit.org/show_bug.cgi?id=286170
Import muprof under Source/ThirdParty/, and arrange CMake to build them
for the GTK and WPE ports. The ENABLE_SYSPROF_MUPROF build option may be
used to toggle compiling it.

Also add libdex, which is a dependency of muprof, with an option to
choose whether to use the system-provided version, or to statically link
the bundled copy (USE_SYSTEM_LIBDEX).

Note that while muprof is GPL-licensed, it does not get linked into
any WebKit library: it is a standalone tool. As for libdex, it is
LGPL-licensed and only linked with muprof, and both licenses are
compatible.

Sysprof is disabled by default. It can be enabled by passing
-DUSE_SYSPROF_CAPTURE=ON to the build, which will add the profiling
marsk to the code.

Enabling sysprof will enable muprof by default as the client to use.
Muprof can be disabled by passing -DENABLE_SYSPROF_MUPROF=OFF to the
build.