Reclaim freed SQLite filesystem space
When data is removed from a SQLite database, pages are marked as free
instead of being release back to the OS / filesystem. Consequently, when
an app uses a lot of IndexedDB space at some point in time, even if the
database contents are deleted afterwards, the database file size will
remain with the max size so far. This can cause filesystem space to get
exhausted in constrained environments where small volume sizes are used
for storage.It can also cause issues with the database operations when
full volume capacity was reached (e.g. multiple domains, local storage
usage) despite not using the full "real" quota and WAL is being used.

These changes attempt to mitigate this issue by running the incremental
auto vacuum whenever some amount of free space is available to be
released. This is currently done on database open. While not ideal, as
there is no guarantee that the database will be opened again to free
that space, there is also no guarantee that the database close method
( SQLiteIDBBackingStore::close() ) will be invoked in all cases. The
cleanup can also not be run in sqlite hooks, as it would not be safe
due potential conflict with ongoing transactions.

The changes also contain a fix for the incremental vacuum operation
itself, as it was not being correctly performed and was returning an
error when it would actually succeed.