[JSC][32-bit] Fix missing move when shift == 0 in BBQ https://bugs.webkit.org/show_bug.cgi?id=300509
Reviewed by Justin Michaud.

In shiftI64Helper, we would check if the shift amount was zero and just jump to
the end, without moving the lhs to the result.

No new tests, but this was found when analysing the crash in JetStream3's
tfjs-wasm.js. 32-bit JSC still crashes when running that test but this gets us
a little bit further.

* Source/JavaScriptCore/wasm/WasmBBQJIT32_64.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::shiftI64Helper):

Canonical link: https://commits.webkit.org/301475@main